style(dupServices): rename duplicate services
